mst: large re-ordering and update from typescript version
Unfortunately this wiped a bunch of edit history, making it hard to see
what actually changed vs just moved. But it was getting difficult to do
these edits without having the code in the same order.
